Fixed many of the "comparison between signed and unsigned integer expressions" warnings.
git-svn-id: file:///svn/uhexen2/code/trunk@4132 eaf649e6-6c13-0410-b194-84221022858e

+++ engine/h2shared/d_fill.c
@@ -33,10 +33,10 @@ void D_FillRect (vrect_t *rect, int color)
 		rheight += ry;
 		ry = 0;
 	}
-	if (rx+rwidth > vid.width)
-		rwidth = vid.width - rx;
-	if (ry+rheight > vid.height)
-		rheight = vid.height - rx;
+	if (rx+rwidth  > (int)vid.width)
+		rwidth = (int)vid.width - rx;
+	if (ry+rheight  > (int)vid.height)
+		rheight = (int)vid.height - rx;
 
 	if (rwidth < 1 || rheight < 1)
 		return;
