+++ engine/hexen2/host_cmd.c
@@ -1075,7 +1075,7 @@ static int LoadGamestate (const char *level, const char *startspot, int ClientsM
 	while (!feof(f))
 	{
 		fscanf (f, "%i\n", &entnum);
-		for (i = 0; i < sizeof(str)-1; i++)
+		for (i = 0; i < (int) sizeof(str) - 1; i++)
 		{
 			r = fgetc (f);
 			if (r == EOF || !r)
@@ -1087,7 +1087,7 @@ static int LoadGamestate (const char *level, const char *startspot, int ClientsM
 				break;
 			}
 		}
-		if (i == sizeof(str)-1)
+		if (i == (int) sizeof(str) - 1)
 			Host_Error ("%s: Loadgame buffer overflow", __thisfunc__);
 		str[i] = 0;
 		start = str;
@@ -1393,8 +1393,8 @@ static void Host_Say (qboolean teamonly)
 		q_snprintf (text, sizeof(text), "\001<%s> %s", hostname.string, p);
 
 // check length & truncate if necessary
-	j = strlen(text);
-	if (j >= sizeof(text) - 1)
+	j = (int) strlen(text);
+	if (j >= (int) sizeof(text) - 1)
 	{
 		text[sizeof(text) - 2] = '\n';
 		text[sizeof(text) - 1] = '\0';
@@ -1471,8 +1471,8 @@ static void Host_Tell_f (void)
 	q_snprintf (text, sizeof(text), "%s: %s", host_client->name, p);
 
 // check length & truncate if necessary
-	j = strlen(text);
-	if (j >= sizeof(text) - 1)
+	j = (int) strlen(text);
+	if (j >= (int) sizeof(text) - 1)
 	{
 		text[sizeof(text) - 2] = '\n';
 		text[sizeof(text) - 1] = '\0';
