+++ engine/hexenworld/server/sys_unix.c
@@ -336,7 +336,7 @@ char *Sys_ConsoleInput (void)
 		}
 		con_text[textlen] = c;
 		textlen++;
-		if (textlen < sizeof(con_text))
+		if (textlen < (int) sizeof(con_text))
 			con_text[textlen] = '\0';
 		else
 		{
